refactor: split api v3 wiring into runtime and override modules
